animations: f32 sub-pixel render_rect to kill ease-out tail judder
ease_out_cubic flattens hard near t=1 — the last 5% of duration covers
~0.0001 of the curve. With 150 ms tweens at 240 Hz that tail is 1–2
frames advancing fractions of a pixel; round-to-int meant a 0-pixel
frame followed by a 1-pixel frame, visible as a small stair-step
right before the window settles.

Introduce RectF (f32 x/y/w/h) alongside Rect (i32) and route every
draw-relevant value through float: Window.render_rect / from_rect,
PlacedBuffer's rect, SceneElement.x/y/dst_width/dst_height, and the
DRM presenter's NDC math. The bilinear filter (already on for both
SHM and dmabuf paths) absorbs the fractional offset into smooth
texture sampling, so the tail glides instead of stepping.

Layout, hit-testing, and focus borders stay on Rect (i32) — they
care about whole pixels and gain nothing from sub-pixel positioning.
RectF::round() / From<Rect> handle the conversion at those few
boundaries.

+++ crates/voidptr-core/src/lib.rs
@@ -50,19 +50,24 @@ pub struct SceneElement<'a> {
     pub buffer_key: u64,
     pub width: i32,
     pub height: i32,
-    /// Top-left in screen coordinates.
-    pub x: i32,
-    pub y: i32,
+    /// Top-left in screen coordinates. Sub-pixel precision (f32)
+    /// because layout-change animations interpolate continuously and
+    /// rounding to int per frame produces visible micro-judder near
+    /// the curve's tail. The presenter feeds these directly into
+    /// the NDC transform; the GPU's bilinear filter absorbs the
+    /// fractional offset.
+    pub x: f32,
+    pub y: f32,
     /// Output rect size (the compositor's chosen on-screen dimensions).
-    /// When `0`, fall back to `width`/`height` (source) — the
+    /// When `<= 0.0`, fall back to `width`/`height` (source) — the
     /// sentinel for subsurfaces / cursor that always render 1:1
     /// at buffer size. Toplevels and layer-shell surfaces pass
     /// the tile / anchor rect so a buffer larger or smaller than
     /// the output rect (common when `wp_viewport.set_source` is
     /// used to crop a buffer down to a sub-rect) gets stretched
     /// or cropped via UV into the intended quad.
-    pub dst_width: i32,
-    pub dst_height: i32,
+    pub dst_width: f32,
+    pub dst_height: f32,
     pub uv_x: f32,
     pub uv_y: f32,
     pub uv_w: f32,
